// src/pages/Senaia/IDENTIFICA/LiquidacionScreen/types.ts
export interface ILiquidacion {
  dbversion?: string;
  liquidacionId?: string;
  liquidacionUniqueId?: string;
  liquidacion_company?: string;
  liquidacionGuias?: string;
  liquidacionCantidad?: string;
  liquidacionFaena_kg_total?: string;
  liquidacionPromedio_animal?: string;
  liquidacionPrecio_venta?: string;
  liquidacionFecha_salida?: string;
  liquidacionFecha_faena?: string;
  liquidacionPrecio_kg_carcasa?: string;
  liquidacionPrecio_total?: string;
}